Criando padrão middleware no backend
