+++ backend/src/config/Firebase.js
@@ -12,16 +12,6 @@ class Firebase {
 			.auth()
 			.createUser({ email: data.email, password: data.password });
 	}
-	async isAuthenticated(req, res, next) {
-		try {
-			const token = req.headers.authorization.split(" ")[1].split('"')[1];
-			const idToken = await admin.auth().verifyIdToken(token);
-			req.decodedToken = idToken;
-			return next();
-		} catch (e) {
-			return res.status(400).json({ error: "Usuário não autorizado" });
-		}
-	}
 }
 
 export default new Firebase();
