+++ backend/src/models/MedicalReport.js
@@ -10,7 +10,8 @@ const MedicalReportSchema = new mongoose.Schema({
 	smoker: Boolean,
 	foodRestriction: [String],
 	allergies: [String],
-	exams: [{ name: String, base64: String, show: Boolean }],
+	exams: [{ name: String, base64: String }],
+
 });
 
 export default mongoose.model("MedicalReport", MedicalReportSchema);
